Finished unit testing on TestAPilot.cs and commenting Attack, Repair and Refuel actions.

+++ Umshini/Battle/BasicBattle.cs
@@ -207,7 +207,7 @@ namespace Battle
                     return false;
                 }
 
-            } while (!currentPilot.Refuel((int)iChoice, lInputActions));
+            } while (!currentPilot.Refuel(iChoice, lInputActions));
 
             lInputActions.Add((int)iChoice);
